Updating console return value handling.
Commands (and therefore the Dispatcher) now return response objects.
A status property has been added to the `Response` class.

If an action returns an integer it is used directly to set the status
code. If it returns a boolean, `false` sets the status code to `1`
whereas `true` sets it to `0`. All other values are compared in a not
strictly way and the result evaluated as with booleans.

Updating tests and mocks.