Refactoring `Catalog`.
Forcing uniform format for in and ouput.

The conversions for the `lossy` and `lossless` have been moved into the
`Catalog` class' `read` and `write` methods.

Removing support for querying multiple locales.

Support for multiple locales is seldomly needed and adds up to the
complexity and speed of the methods.

Removing explitcly defined `Catalog` categories.

To allow for easier extension and adding of categories checks if a
certain category is supported have been moved into the `read` and
`write` methods of the adapters.

Renaming categories for templates and pages.

By default all categories are treated as socalled _pages_. In order for
being able to treat all categories in a uniform way. Adapters only have
to return full pages and don't need to care about it's components and
support for them.

Selecting of _page_ ids.

Parts of _pages_ can be selected by appending the id to select to the
category: `validation.phone`.

Renaming and making `Base::mergeMessageItem()` more generic.

Renaming `Base::formatMessageItem()`.

Upating lossless format to be as flexible as possible.