updates to the 'li3 library extract' command
Added setting the namespace in files and the default library name according to the basename of the directory being extracted to. Can be explicitly set with the --namespace option.

Also updated fix for http://dev.lithify.me/lithium/tickets/view/32 to not hardcode the `LITHIUM_LIBRARY_PATH` if it is the same as the default `dirname(LITHIUM_APP_PATH) . '/libraries'` value which is more portable.

Lastly, library archives can include a special _replacements.json file that will override the default replacements that happen when a library is extracted.