fixing controller rendering of requests with a set type
If you set the request type manually in the dispatch cycle, it was not being honored by `Controller` which would always use text/html.  This change fixes that.  It was found that it broke form requests, so the `Media` class type config was updated to include `responseType` as an option.  This is currently only needed for `form` requests where the incoming type is 'form' but the response type needs to be 'html'.  Integration and unit tests were added.