Disabling memcache extension installation on CI.
This forces tests to skip. Although the CI environment was featuring a
running Memcached it wasn't advertised and sometimes not working as
expected, causing tests to fail.

This patch can be reverted when the CI environment starts to advertise a
Memcached or such a daemon can be installed and started manually.