suppressEolIssues
On linux EOL is LF, on Windows it is normally CRLF, but the latter may depend also
on the git config core.autocrlf setting. As some tests use heredoc style (<<<) to
specify multiline expectations, this EOL issue may cause tests to fail only because
of a difference in EOL's used.

in assertEqual, assertNotEqual, assertPattern and assertNotPattern this function is
called to get rid of any EOL differences.