Add `setupFunctions()` method to `Mocker`.
This allows you to declare which methods you are going to define later on. This, for instance, may be helpful if something like apc is caching the full namespace of a method. The first time you call a global function, it might be cached, so stubbing it would not work.