Update `Media` to accept an array of `host` to be more dynamic.
For a site that uses a CDN, but has an uneven amount of js/css/imgages. They might upload all assets to all CDN's then have a script to 'randomly' select one based on the filename. Doing it based on the filename ensures the same CDN will always get picked, which allows for browser caching.

To be applied on top of #878.