Fix hasMany record set returned from `with` key on a finder.
The record set built from records read from a hasMany relationship should use the related model, not the original model. If the original model is used, the record set is indexed by the original model's key. Since this is a hasMany relation, the original model's key is set on each related record, and thus all related records will share the same index in the record set.