Adding support for HTTP basic/digest auth for CGI environments in `Request`.
Adding tests for `lithium\action\Request`.

This fixes basic/digest auth for CGI environments in
`lithium\security\auth\adapter\Http`.

Parts of the implementation (the decoding of the basic auth header)
could in the future be extracted into a new decode method in
`lithium\net\http\Auth`. However there are two arguments against that:

(a) `lithium\net\http\Auth::decode` currently works expects a digest
    header only - with the leading `"Digest "` part already stripped off. To
    make the method authorization method agnostic it would need to expect
    the unstripped header or another hint in form of an additional argument
    to the method.

(b) Even if the above is implemented we would create a dependency on
    `Auth` which may not be desired or needs to be further clarified.