Moving filtering from cache adapters into manager.
Functionality should not be needed to repeated in adapters and
handled by the manager class if possible.

Also adapter methods themselves were never filterable on the
adapter classes. It was always (parts of) the manager class'
methods that were filterable.

That's why filters already in use with the cache manager class
will continue to work as is.

- Updating docblocks.
- Updating and cleaning up all relevant tests.
- Make read/write/delete/increment/decrement methods to not
  return closures anymore.

BC Break: This change requires any cache adapters implemented in
userland to not return closures for read/write/delete/increment/decrement
anymore. Please update your code accordingly.

When using the standard lithium adapters no change ist required
any code will continue to work as is.