Providing soft deprecation when directly loading `Apc` adapter file.
All cache adapter extend a base cache adapter class since
7b4de1d06a46f45e95ec2d4d5bdcea5776fd491c.

The apc adapter was manually loaded in the app bootstrap (which has
already been changed not to do so anymore) in order to optimize the
startup time.

It can be assumed that users may upgrade their lithium installation
but not the app. Until they do this commit provides a workaround.

This deprecation notice should be removed with the next minor version
release.