Refactor model finders and document find.
Removes treatment of `'all'` finder as a special case, by
actually adding it as a finder. Normalize all finders as
closures when being added using `finder()`.

These changes reduces the amount of code
in `find()` and cases in `__callStatic()`.

Renamed to `_finders()` for clarity. Finders are a special class
of filters. Users may not know about filters, but know of finders.

Brings `_finders()` method closer to `find()` and `finder()`, where
it is actually used.

Adds thorough API documentation.

Makes method static, so we can later directly call, without using $self.