Fix failing test; fully normalize filters in test dispatcher.
Report filters have been refactored to be more easy to understand
(by removing lazy init, which is not required in this case).

The setter/getter has been documented, it expects filter options
to be arrays.

Test filters are now fully normalized (values are now always an
array) when coming from test dispatcher.